requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary
pydantic>=2.0
python-dotenv
passlib[bcrypt]
django
djangorestframework
pgvector
